Use progress bar to display scraping progress
Unfortunately, this isn't 100% accurate - the progress bar is
incremented for a crate before all of its versions have been fetched.

Rayon provides no way of running callbacks when scoped tasks complete,
and indicatif doesn't allow us to atomically increase the progress bar
length. Unless either of these crates expands their API, this is the
best we can do.

In practice, this doesn't end up mattering that much - it manifests
itself as the progressbar briefly hanging at 100%. Unless we're very
unlucky, most of the version-fetcing tasks should have run by this
point, so we won't be waiting for very long.