Handle our own consolidation
Upstream's appears to be buggy. For example `cargo tally syn-test-suite`
returns 0 results with consolidate and 1 result without consolidate,
with 1 being the correct answer, even though consolidate is documented
as having no semantic effect.

Our handling of the result collection already performs the equivalent of
consolidation so consolidating as part of the dataflow is redundant.